Refactor for X25519; Use X25519_keypair
